refactor : 폴더 구조 변경, 매칭 관련 기능은 하나의 컨트롤러에서 수행하도록 수정
