--- src/main/java/com/_z/eum/matching/matchResult/dto/request/MatchRequest.java
@@ -1,7 +1,0 @@
-package com._z.eum.matching.matchResult.dto.request;
-
-import java.util.List;
-
-public record MatchRequest(
-        List<Integer> optionIds
-) {}
