+++ src/main/java/com/_z/eum/matching/repository/CareerTestRepository.java
@@ -1,12 +1,11 @@
-package com._z.eum.matching.careerTest.repository;
+package com._z.eum.matching.repository;
 
-import com._z.eum.matching.careerTest.entity.CareerTestQuestion;
+import com._z.eum.matching.entity.CareerTestQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface CareerTestRepository extends JpaRepository<CareerTestQuestion,Integer> {
